# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triangle Curling Spares API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/leagues": {
      "get": {
        "tags": [
          "leagues"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "dayOfWeek": {
                        "type": "number"
                      },
                      "format": {
                        "type": "string",
                        "enum": [
                          "teams",
                          "doubles"
                        ]
                      },
                      "startDate": {
                        "type": "string"
                      },
                      "endDate": {
                        "type": "string"
                      },
                      "drawTimes": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "exceptions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "dayOfWeek",
                      "format",
                      "startDate",
                      "endDate",
                      "drawTimes",
                      "exceptions"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "leagues"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "dayOfWeek": {
                    "type": "number"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "teams",
                      "doubles"
                    ]
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "endDate": {
                    "type": "string"
                  },
                  "drawTimes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "exceptions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name",
                  "dayOfWeek",
                  "format",
                  "startDate",
                  "endDate",
                  "drawTimes"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "dayOfWeek": {
                      "type": "number"
                    },
                    "format": {
                      "type": "string",
                      "enum": [
                        "teams",
                        "doubles"
                      ]
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "endDate": {
                      "type": "string"
                    },
                    "drawTimes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "exceptions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "dayOfWeek",
                    "format",
                    "startDate",
                    "endDate",
                    "drawTimes",
                    "exceptions"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/leagues/{id}/upcoming-games": {
      "get": {
        "tags": [
          "leagues"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "date": {
                        "type": "string"
                      },
                      "time": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "date",
                      "time"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}